src/agents/luotianyi_agent.py
"""
洛天依Agent主类

实现洛天依角色扮演对话Agent的核心逻辑
"""

from typing import Dict, List, Optional, Any
from abc import ABC, abstractmethod
import yaml
import os

from ..llm.siliconflow_client import SiliconFlowClient
from ..knowledge.graph_retriever import GraphRetriever
from ..knowledge.vector_store import VectorStore
from ..llm.prompt_manager import PromptManager
from .conversation_manager import ConversationManager
from ..utils.logger import get_logger


class BaseAgent(ABC):
    """Agent基类"""
    
    @abstractmethod
    def chat(self, message: str, **kwargs) -> str:
        """处理用户消息并返回回复"""
        pass
    
    @abstractmethod
    def reset(self) -> None:
        """重置对话状态"""
        pass


class LuoTianyiAgent(BaseAgent):
    """洛天依对话Agent
    
    基于图结构增强RAG的洛天依角色扮演对话系统
    """
    
    def __init__(self, config_path: str = "config/config.yaml"):
        """初始化洛天依Agent
        
        Args:
            config_path: 配置文件路径
        """
        self.logger = get_logger(__name__)
        self.config = self._load_config(config_path)
        self.persona = self._load_persona()
        
        # 初始化各个组件
        self.llm_client = self._init_llm_client()
        self.vector_store = self._init_vector_store()
        self.graph_retriever = self._init_graph_retriever()
        self.prompt_manager = self._init_prompt_manager()
        self.conversation_manager = self._init_conversation_manager()
        
        self.logger.info("洛天依Agent初始化完成")
    
    def _load_config(self, config_path: str) -> Dict[str, Any]:
        """加载配置文件
        
        Args:
            config_path: 配置文件路径
            
        Returns:
            配置字典
        """
        # TODO: 实现配置文件加载逻辑
        # - 读取YAML配置文件
        # - 处理环境变量替换
        # - 验证配置格式
        pass
    
    def _load_persona(self) -> Dict[str, Any]:
        """加载人设配置
        
        Returns:
            人设配置字典
        """
        # TODO: 实现人设配置加载
        # - 从配置文件中读取persona_file路径
        # - 加载洛天依人设YAML文件
        # - 解析人设信息
        pass
    
    def _init_llm_client(self) -> SiliconFlowClient:
        """初始化LLM客户端
        
        Returns:
            LLM客户端实例
        """
        # TODO: 根据配置初始化硅基流动客户端
        pass
    
    def _init_vector_store(self) -> VectorStore:
        """初始化向量存储
        
        Returns:
            向量存储实例
        """
        # TODO: 根据配置初始化向量数据库
        pass
    
    def _init_graph_retriever(self) -> GraphRetriever:
        """初始化图检索器
        
        Returns:
            图检索器实例
        """
        # TODO: 根据配置初始化图数据库检索
        pass
    
    def _init_prompt_manager(self) -> PromptManager:
        """初始化Prompt管理器
        
        Returns:
            Prompt管理器实例
        """
        # TODO: 初始化Prompt模板管理
        pass
    
    def _init_conversation_manager(self) -> ConversationManager:
        """初始化对话管理器
        
        Returns:
            对话管理器实例
        """
        # TODO: 初始化对话历史和上下文管理
        pass
    
    def chat(self, message: str, **kwargs) -> str:
        """处理用户消息并生成洛天依风格的回复
        
        Args:
            message: 用户输入消息
            **kwargs: 额外参数
            
        Returns:
            洛天依的回复文本
        """
        try:
            self.logger.info(f"接收到用户消息: {message}")
            
            # 1. 意图理解和实体识别
            intent_result = self._understand_intent(message)
            
            # 2. 知识检索
            retrieved_knowledge = self._retrieve_knowledge(message, intent_result)
            
            # 3. 生成回复
            response = self._generate_response(message, intent_result, retrieved_knowledge)
            
            # 4. 更新对话历史
            self._update_conversation_history(message, response)
            
            self.logger.info(f"生成回复: {response}")
            return response
            
        except Exception as e:
            self.logger.error(f"对话处理出错: {e}")
            return self._get_error_response()
    
    def _understand_intent(self, message: str) -> Dict[str, Any]:
        """理解用户意图和识别实体
        
        Args:
            message: 用户消息
            
        Returns:
            意图分析结果
        """
        # TODO: 实现意图理解逻辑
        # - 使用NLP工具进行分词和实体识别
        # - 分类用户意图（问候、询问歌曲、日常聊天等）
        # - 提取关键实体信息
        pass
    
    def _retrieve_knowledge(self, message: str, intent_result: Dict[str, Any]) -> Dict[str, Any]:
        """检索相关知识
        
        Args:
            message: 用户消息
            intent_result: 意图分析结果
            
        Returns:
            检索到的知识信息
        """
        # TODO: 实现知识检索逻辑
        # - 根据意图和实体进行向量检索
        # - 使用图结构进行多跳推理
        # - 合并和排序检索结果
        pass
    
    def _generate_response(self, message: str, intent_result: Dict[str, Any], 
                          knowledge: Dict[str, Any]) -> str:
        """生成洛天依风格的回复
        
        Args:
            message: 用户消息
            intent_result: 意图分析结果
            knowledge: 检索到的知识
            
        Returns:
            生成的回复文本
        """
        # TODO: 实现回复生成逻辑
        # - 构建包含人设和知识的Prompt
        # - 调用LLM生成回复
        # - 后处理确保风格一致性
        pass
    
    def _update_conversation_history(self, user_message: str, bot_response: str) -> None:
        """更新对话历史
        
        Args:
            user_message: 用户消息
            bot_response: 机器人回复
        """
        # TODO: 实现对话历史更新
        pass
    
    def _get_error_response(self) -> str:
        """获取错误回复
        
        Returns:
            错误情况下的默认回复
        """
        # TODO: 根据人设返回合适的错误回复
        return "诶？天依有点不太明白呢～可以再说一遍吗？"
    
    def reset(self) -> None:
        """重置对话状态"""
        # TODO: 实现状态重置
        # - 清空对话历史
        # - 重置上下文状态
        self.logger.info("对话状态已重置")
    
    def get_conversation_history(self) -> List[Dict[str, str]]:
        """获取对话历史
        
        Returns:
            对话历史列表
        """
        # TODO: 返回对话历史
        pass
    
    def update_persona(self, persona_updates: Dict[str, Any]) -> None:
        """更新人设配置
        
        Args:
            persona_updates: 人设更新内容
        """
        # TODO: 实现人设动态更新
        pass
